pyproject.toml
[project]
name = "ollama-web"
version = "0.2.0"
description = "A web application for analyzing images and documents using Ollama models"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "flask>=3.0.0",
    "flask-wtf>=1.2.1",
    "werkzeug>=3.0.0",
    "ollama>=0.1.6",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.6.1",
]
requires-python = ">=3.10"

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.rye]
managed = true
dependencies = [
    "flask>=3.0.0",
    "flask-wtf>=1.2.1",
    "werkzeug>=3.0.0",
    "ollama>=0.1.6",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.3.1",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
pythonpath = [
    "."
]
